ClaudeHTTPS_FixedV1/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
paho-mqtt==2.1.0
pydantic==2.9.0
python-dotenv==1.0.1
